% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_company.R
\name{analyze_company}
\alias{analyze_company}
\title{Analyze Company Information}
\usage{
analyze_company(company1, company2)
}
\arguments{
\item{company1}{character string representing the first company name.}

\item{company2}{character string representing the second company name.}
}
\value{
This function does not return a value, but creates three plots.
}
\description{
This function takes the name of two companies and analyzes their information based on data from a movie dataset.
}
\examples{
analyze_company("Paramount Pictures","Universal Pictures")

}
