% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emotion_score.R
\name{emotion_score}
\alias{emotion_score}
\title{Calculate Emotion Scores}
\usage{
emotion_score(data)
}
\arguments{
\item{data}{A character string containing the input text to analyze}
}
\value{
A bar chart showing the emotion scores for each category and a string describing the two most prevalent emotions in the input text.
}
\description{
This function calculates emotion scores for a given text using the NRC sentiment lexicon and plots a bar chart showing the scores for each emotion category.
}
\examples{
emotion_score("jaunty Hit me again, Sven. The moment of truth boys. Somebody's life's about to change. Let's see... Fabrizio's got niente. Olaf, you've got squat. Sven, uh oh... two pair... mmm. turns to his friend Sorry Fabrizio. Sorry, you're not gonna see your mama again for a long time... grinning 'Cause you're goin' to America!! Full house boys! to the Swedes Sorry boys. Three of a kind and a pair. I'm high and you're dry and... to Fabrizio we're going to -- Goin' home... to the land o' the free and the home of the real hot- dogs! On the TITANIC!! We're ridin' in high style now! We're practically goddamned royalty, ragazzo mio!! Shit!! Come on, Fabri! grabbing their stuff Come on!!")
}
