% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_world_map.R
\name{heat_world_map}
\alias{heat_world_map}
\title{Create a plotly map of movie data by decade and type}
\usage{
heat_world_map(decade, type)
}
\arguments{
\item{decade}{A character string representing the decade to plot, which must be one of "1980-1989", "1990-1999", "2000-2009", or "2010-2020".}

\item{type}{A character string representing the type of plot to generate, which must be one of "count", "score", "count_prop", or "score_ranking".}
}
\value{
A plotly map of the specified movie data.
}
\description{
This function generates a map of movie data for a given decade and type.
}
\examples{
# Generate a count map for the 1980s
heat_world_map("1980-1989", "count")

# Generate a score map for the 2010s
heat_world_map("2010-2020", "score")

}
\seealso{
\code{\link{plotly}}

\code{\link{read.csv}}

\code{\link{table}}

\code{\link{mean}}
}
