% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_data_visualization.R
\name{individual_data_visualization}
\alias{individual_data_visualization}
\title{Create a wordcloud and 2 by 2 conbined plot for a specific movie star, director, company, country.}
\usage{
individual_data_visualization(type = "country", name = "China")
}
\arguments{
\item{type}{A character string indicating the type of input. It can be "movie star", "director", "company", "country". Default is "country".}

\item{name}{A character string indicating the name of the movie star, director, company, or country. Default is "China".}
}
\description{
This function generates a wordcloud for the genres of movies produced by a movie star, director, company, or country.
In the 2 by 2 combined plot, it plots one pie chart and one pie circle chart to show the different genres' proportions in the first line; and two line charts to show the total number of movie and the average movie score variations over different years in the second line.
}
\examples{
individual_data_visualization("12345", "Canada")
individual_data_visualization("director", "Stephen Curry")
individual_data_visualization("star", "Denzel Washington")
individual_data_visualization("director", "Steven Spielberg")
individual_data_visualization("company", "Warner Bros.")
individual_data_visualization("country", "Canada")

}
