% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Jack}
\alias{Jack}
\title{Jack}
\format{
An object of class \code{data.frame} with 1 rows and 1 columns.
}
\usage{
data(Jack)
}
\description{
This dataset contains all of Jack's lines from Titanic, and can be used to do emotion analysis later.
}
\keyword{datasets}
