% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{movie_data}
\alias{movie_data}
\title{Movies Data}
\format{
An object of class \code{data.frame} with 7668 rows and 15 columns.
}
\usage{
data(movie_data)
}
\description{
This dataset contains the information of 53940 movies with 14 attributes.
Each row in the table represents a different movie, and the following are the details of each column:
}
\details{
\itemize{
  \item \code{rating}: The rating of the movie (e.g., G, PG, R).
  \item \code{genre}: The genre of the movie (e.g., Drama, Adventure, Comedy).
  \item \code{year}: The year the movie was released.
  \item \code{released}: The date the movie was released (including the country).
  \item \code{score}: The average rating of the movie (out of 10).
  \item \code{votes}: The number of votes used to calculate the movie's score.
  \item \code{director}: The name of the movie's director.
  \item \code{writer}: The name of the movie's writer.
  \item \code{star}: The name of the movie's star (lead actor/actress).
  \item \code{country}: The country where the movie was produced.
  \item \code{budget}: The production budget of the movie.
  \item \code{gross}: The box office gross of the movie.
  \item \code{company}: The production company that made the movie.
  \item \code{runtime}: The length of the movie in minutes.
}
}
\keyword{datasets}
