% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reco.R
\name{reco}
\alias{reco}
\title{Calculate recommendations based on user preferences}
\usage{
reco(history, method, id)
}
\arguments{
\item{history}{A data frame containing columns, namely, "user.id", "movie", and "rating".}

\item{method}{Algorithms used in predicting user preferences, such as "UBCF","IBCF","RANDOM","POPULAR","SVD".}

\item{id}{The id of the user whom we need to recommend to.}
}
\value{
A list of top3 recommended items, and a measure of performance evaluation parameters such as MAE,MSE,and RMSE.
}
\description{
This function takes in a user's preferences and generates recommendations for items
that match those preferences.
}
\examples{
data(movie_data)
input <- data.frame("user.id"=sample(1:50, 6000,replace = TRUE), "movie"=sample(movie_data$name,6000), "rating"=sample(1:10, 6000,replace = TRUE))
reco(input,"IBCF",1)

}
