% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Rose}
\alias{Rose}
\title{Rose}
\format{
An object of class \code{data.frame} with 1 rows and 2302 columns.
}
\usage{
data(Rose)
}
\description{
This dataset contains all of Rose's lines from Titanic, and can be used to do emotion analysis later.
}
\keyword{datasets}
