% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_predict.R
\name{score_predict}
\alias{score_predict}
\title{Predict movie scores using different machine learning models}
\usage{
score_predict(method, newdata)
}
\arguments{
\item{method}{A character string indicating the machine learning method to use. The available options are "svm", "neural network", "decision tree", "randomforest", and "linear regression".}

\item{newdata}{A data frame containing the features of budget, runtime, votes, gross in order.}
}
\value{
The predicted score of the new movie based on the chosen machine learning method.
}
\description{
This function takes in a machine learning method and new data and returns a predicted score using that method.
}
\examples{
score_predict("linear regression", newdata = data.frame(budget = 10000000, runtime = 120, votes = 2000, gross = 1000000))
score_predict("neural network", newdata = data.frame(budget = 12000000, runtime = 106, votes = 4000, gross = 1650000))
}
