% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_movies_comparison.R
\name{top_movies_comparison}
\alias{top_movies_comparison}
\title{Create a 2 by 2 conbined plot for the top movies.}
\usage{
top_movies_comparison(
  time.start = 1900,
  time.end = 2022,
  sort.criterion = "score",
  show.number = 10
)
}
\arguments{
\item{time.start}{the start year of the movie data to be analyzed (default: 1900)}

\item{time.end}{the end year of the movie data to be analyzed (default: 2022)}

\item{sort.criterion}{the criterion by which to sort the movie data ("score" or "gross", default: "score")}

\item{show.number}{the number of movies to display, suggest 10 or 20 (default: 10)}
}
\value{
This function has no return value, but prints out the filtered and sorted movie data as well as four bar charts of the count of movies by rating, genre, company, and country.
}
\description{
top_movies_comparison targets on the comparison between a specific number of the top movies under a specific sorting criterion("score" or "gross"). It will generate one 2 by 2 conbined plot.
}
\examples{
top_movies_comparison(1900, 2020, "score", 20)
top_movies_comparison(2000, 2020, "gross", 10)

}
