% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_similarity.R
\name{user_similarity}
\alias{user_similarity}
\title{Calculate user similarity matrix based on user's historical movie genre ratings}
\usage{
user_similarity(history)
}
\arguments{
\item{history}{A data frame containing columns, namely, "user.id", "movie", and "rating".}
}
\description{
This function takes a data frame which is composed of 'user.id','movie', and 'ratings' and calculates a user similarity matrix based on the genre ratings of the movies. The similarity matrix is calculated using cosine similarity.
}
\examples{
input <- data.frame("user.id"=sample(1:10, 5000,replace = TRUE), "movie"=sample(movie_data$name,5000), "rating"=sample(1:10, 5000,replace = TRUE))
user_similarity(input)

}
